(program.cpp) Add call to input_poll before input_state
The libretro spec says the input_poll callback must be called at least once a frame. The problem is because this core was tested with RetroArch's default polling type. With the default type, "Late", the frontend itself ensures input_poll is called only once on the first input_state call. 

But, because this polling type also ignores the core's input_poll calls, RetroArch makes sure to call input_poll after running the core for 1 frame if input_state was never called, so as not to break the spec. And this masked the problem.